Add a few flex points to the JSONWebTokenBackend
This moves all loading of users to `get_user()` and expands it to taking
kwargs. Developers can override the values that used for finding the
user by overriding `get_user_kwargs`. Note: if a `user_id` is provided
to `get_user`, it is translated to `pk` and will overwrite any `pk`
provided.